+++ monarchmoney/__init__.py
@@ -12,5 +12,5 @@ from .monarchmoney import (
     RequireMFAException,
 )
 
-__version__ = "0.2.1"
+__version__ = "0.2.2"
 __author__ = "keithah"